skip gh-pages

+++ .build.sh
@@ -10,15 +10,18 @@ fi
 mkdir dist
 
 for tag in `git ls-remote --heads origin  | sed 's?.*refs/heads/??'`; do
-  mkdir dist/$tag
-  git reset --hard origin/$tag
-  echo "Building $tag"
-  touch _book && rm -r _book
-  gitbook install > /dev/null
-  gitbook build > /dev/null
-  cp -r _book/* dist/$tag/
+  if [ "$tag" -ne "gh-pages" ]; then
+    mkdir dist/$tag
+    git reset --hard origin/$tag
+    echo "Building $tag"
+    touch _book && rm -r _book
+    gitbook install > /dev/null
+    gitbook build > /dev/null
+    cp -r _book/* dist/$tag/
+  fi
 done;
 
+"
 for tag in `git tag`; do
   mkdir dist/$tag
   git reset --hard $tag
@@ -29,9 +32,11 @@ for tag in `git tag`; do
   cp -r _book/* dist/$tag/
 done;
 
-git reset --hard $latest
 echo "Building $tag"
 touch _book && rm -r _book
 gitbook install > /dev/null
 gitbook build > /dev/null
 cp -r _book/* dist/
+"
+
+git reset --hard $latest
